+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added support for Realtime Database Auth Variable Overrides
   ([#625](https://github.com/kreait/firebase-php/pull/625))
   ([Documentation](https://firebase-php.readthedocs.io/en/latest/realtime-database.html#authenticate-with-limited-privileges))
+* Added support for linking IdP credentials to an existing account
+  ([#635](https://github.com/kreait/firebase-php/pull/635))
+  ([Documentation](https://firebase-php.readthedocs.io/en/latest/authentication.html#linking-and-unlinking-idp))
 ### Changes
 * Database Rules are now uploaded as pretty printed JSON to improve readability when viewing
   them in the Firebase Console.